Better visualizations of the list changes

+++ app/src/main/java/com/genonbeta/TrebleShot/activity/ShareActivity.java
@@ -269,7 +269,7 @@ public class ShareActivity extends Activity implements SnackbarSupport
 
 											getProgressDialog().setSecondaryProgress(getProgressDialog().getSecondaryProgress() + 1);
 
-											int requestId = AppUtils.getUniqueNumber();
+											long requestId = AppUtils.getUniqueNumber();
 											JSONObject thisJson = new JSONObject();
 
 											TransferObject transferObject = new TransferObject(requestId,
